Remove bignum_new.c, moving contents to bignum_core.c
