+++ include/mbedtls/mbedtls_config.h
@@ -2016,7 +2016,6 @@
  *          library/bignum_core.c
  *          library/bignum_mod.c
  *          library/bignum_mod_raw.c
- *          library/bignum_new.c
  * Caller:  library/dhm.c
  *          library/ecp.c
  *          library/ecdsa.c
